requirements.txt
web3>=6.0.0
py-solc-x
eth-account>=0.10.0
aiohttp>=3.9.0
colorama>=0.4.6
python-dotenv>=1.0.1
loguru